requirements.txt
azure-functions
azure-storage-blob
pandas
pyodbc
xlrd
